+++ coverage.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" ?>
-<coverage version="6.3" timestamp="1643428131533" lines-valid="266" lines-covered="266" line-rate="1" branches-covered="0" branches-valid="0" branch-rate="0" complexity="0">
+<coverage version="6.3" timestamp="1643431649338" lines-valid="266" lines-covered="266" line-rate="1" branches-covered="0" branches-valid="0" branch-rate="0" complexity="0">
 	<!-- Generated by coverage.py: https://coverage.readthedocs.io -->
 	<!-- Based on https://raw.githubusercontent.com/cobertura/web/master/htdocs/xml/coverage-04.dtd -->
 	<sources>
 		<source>/github/workspace</source>
 	</sources>
 	<packages>
-		<package name="devsetgo_lib" line-rate="1" branch-rate="0" complexity="0">
+		<package name="dsg_lib" line-rate="1" branch-rate="0" complexity="0">
 			<classes>
-				<class name="__init__.py" filename="devsetgo_lib/__init__.py" complexity="0" line-rate="1" branch-rate="0">
+				<class name="__init__.py" filename="dsg_lib/__init__.py" complexity="0" line-rate="1" branch-rate="0">
 					<methods/>
 					<lines/>
 				</class>
-				<class name="calendar_functions.py" filename="devsetgo_lib/calendar_functions.py" complexity="0" line-rate="1" branch-rate="0">
+				<class name="calendar_functions.py" filename="dsg_lib/calendar_functions.py" complexity="0" line-rate="1" branch-rate="0">
 					<methods/>
 					<lines>
 						<line number="2" hits="1"/>
@@ -20,7 +20,7 @@
 						<line number="18" hits="1"/>
 					</lines>
 				</class>
-				<class name="file_functions.py" filename="devsetgo_lib/file_functions.py" complexity="0" line-rate="1" branch-rate="0">
+				<class name="file_functions.py" filename="dsg_lib/file_functions.py" complexity="0" line-rate="1" branch-rate="0">
 					<methods/>
 					<lines>
 						<line number="2" hits="1"/>
@@ -215,7 +215,7 @@
 						<line number="374" hits="1"/>
 					</lines>
 				</class>
-				<class name="folder_functions.py" filename="devsetgo_lib/folder_functions.py" complexity="0" line-rate="1" branch-rate="0">
+				<class name="folder_functions.py" filename="dsg_lib/folder_functions.py" complexity="0" line-rate="1" branch-rate="0">
 					<methods/>
 					<lines>
 						<line number="2" hits="1"/>
@@ -260,7 +260,7 @@
 						<line number="65" hits="1"/>
 					</lines>
 				</class>
-				<class name="logging_config.py" filename="devsetgo_lib/logging_config.py" complexity="0" line-rate="1" branch-rate="0">
+				<class name="logging_config.py" filename="dsg_lib/logging_config.py" complexity="0" line-rate="1" branch-rate="0">
 					<methods/>
 					<lines>
 						<line number="2" hits="1"/>
@@ -276,10 +276,10 @@
 						<line number="56" hits="1"/>
 						<line number="59" hits="1"/>
 						<line number="62" hits="1"/>
-						<line number="101" hits="1"/>
+						<line number="103" hits="1"/>
 					</lines>
 				</class>
-				<class name="patterns.py" filename="devsetgo_lib/patterns.py" complexity="0" line-rate="1" branch-rate="0">
+				<class name="patterns.py" filename="dsg_lib/patterns.py" complexity="0" line-rate="1" branch-rate="0">
 					<methods/>
 					<lines>
 						<line number="2" hits="1"/>
