+++ dsg_lib/patterns.py
