+++ examples/cal_example.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from dsg.calendar_functions import get_month
+from dsg_lib.calendar_functions import get_month
 
 month_list: list = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]
