+++ examples/csv_example.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from dsg.file_functions import open_csv
-from dsg.file_functions import save_csv
+from dsg_lib.file_functions import open_csv
+from dsg_lib.file_functions import save_csv
 
 example_list = [
     ["thing_one", "thing_two"],
