+++ examples/log_example.py
@@ -3,7 +3,7 @@ import logging
 
 from loguru import logger
 
-from dsg import logging_config
+from dsg_lib import logging_config
 
 logging_config.config_log(
     logging_directory="myLoggingFolder",
