+++ examples/pattern_example.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import pprint
 
-from dsg.patterns import pattern_between_two_char
+from dsg_lib.patterns import pattern_between_two_char
 
 ASCII_LIST = [
     " ",
