+++ requirements.txt
@@ -1,20 +1,21 @@
 autoflake==1.4
 autopep8==1.6.0
-black==21.11b1
+black==21.12b0 # from 21.11b1
 bump2version==1.0.1
 Click==8.0.3
+# coverage==6.3
 coverage-badge==1.1.0
 flake8==4.0.1
 loguru==0.5.3
 mkdocs==1.2.3
-mkdocs-material==8.0.4 # from 8.0.2
-pre-commit==2.16.0 # from 2.15.0
-pylint==2.12.2 # from 2.12.1
+mkdocs-material==8.1.8 # from 8.0.4
+pre-commit==2.17.0 # from 2.16.0
+pylint==2.12.2
 pytest==6.2.5
 pytest-cov==3.0.0
 pytest-runner==5.3.1
-tox==3.24.4
+tox==3.24.5 # from 3.24.4
 tqdm==4.62.3
-twine==3.7.0 # from 3.6.0
+twine==3.7.1 # from 3.7.0
 watchdog==2.1.6
-wheel==0.37.0
+wheel==0.37.1 # from 0.37.0
