+++ setup.cfg
@@ -33,8 +33,6 @@ exclude_lines =
 
 [tool:pytest]
 norecursedirs = /tests
-exclude_lines =
-	pragma: no cover
 testpaths = tests
 python_files = test_*.py
 addopts =
