+++ tests/test_file_functions_process.py
@@ -6,10 +6,10 @@ import unittest
 import pytest
 
 # from devsetgo_lib.file_functions import get_data_directory_list
-from dsg.file_functions import create_sample_files
-from dsg.file_functions import open_csv
-from dsg.file_functions import open_json
-from dsg.file_functions import open_text
+from dsg_lib.file_functions import create_sample_files
+from dsg_lib.file_functions import open_csv
+from dsg_lib.file_functions import open_json
+from dsg_lib.file_functions import open_text
 
 time_str = datetime.datetime.now()
