+++ tests/test_file_functions_zprocess_delete.py
@@ -5,7 +5,7 @@ import unittest
 
 import pytest
 
-from dsg.file_functions import delete_file
+from dsg_lib.file_functions import delete_file
 
 time_str = datetime.datetime.now()
