+++ tests/test_logging_config.py
@@ -5,9 +5,8 @@ import unittest
 
 import pytest
 from _pytest.logging import caplog
-from dsg.logging_config import config_log
+from dsg_lib.logging_config import config_log
 import logging
-
 
 
 def some_func(var1, var2):
@@ -19,7 +18,6 @@ def some_func(var1, var2):
 
     return var1 + var2
 
-    
 
 # class Test(unittest.TestCase):
 # `some_func` adds two numbers, and logs a warning if the first is < 1
@@ -28,8 +26,9 @@ def test_some_func_logs_warning(caplog):
     assert some_func(-1, 3) == 2
     assert "Oh no!" in caplog.text
 
+
 def test_exit():
-    
+
     with pytest.raises(SystemExit) as e:
         # The command to test
         config_log(logging_level="bob")
