+++ tests/test_zfolder_functions.py
@@ -9,10 +9,10 @@ from pathlib import Path
 
 import pytest
 
-from dsg.folder_functions import get_directory_list
-from dsg.folder_functions import last_data_files_changed
-from dsg.folder_functions import make_folder
-from dsg.folder_functions import remove_folder
+from dsg_lib.folder_functions import get_directory_list
+from dsg_lib.folder_functions import last_data_files_changed
+from dsg_lib.folder_functions import make_folder
+from dsg_lib.folder_functions import remove_folder
 
 time_str = datetime.now()
